+++ packages/spacecat-shared-data-access/src/v2/models/organization/organization.model.js
@@ -10,7 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-import { Config } from '../../../models/site/config.js';
 import BaseModel from '../base/base.model.js';
 
 /**
@@ -23,9 +22,6 @@ import BaseModel from '../base/base.model.js';
 class Organization extends BaseModel {
   // add your custom methods or overrides here
 
-  getConfig() {
-    return Config(this.record.config);
-  }
 }
 
 export default Organization;